Implement custom hook for media fetching and simplify Home component
